monarch_feeder/computer_use_demo/requirements.txt
streamlit==1.41.0
anthropic[bedrock,vertex]>=0.39.0
jsonschema==4.22.0
boto3>=1.28.57
google-auth<3,>=2
jinja2==3.1.6
python-dotenv==1.1.0
instructor==1.8.3